API_Endpoint_Labelee.py
from flask import Flask, request, jsonify
from Labelee import Labelee  # Import your Labelee model class
import torch
from PIL import Image
import io

app = Flask(__name__)
model = Labelee(vocab_size=10000, feature_dim=768, num_classes=1000)
model.load_state_dict(torch.load("path/to/checkpoint.pth"))  # Load trained weights
model.eval()

@app.route("/api/process", methods=["POST"])
def process_data():
    task = request.form.get("task")
    image = request.files.get("image")
    text = request.form.get("text")

    # Process image
    if image:
        img = Image.open(io.BytesIO(image.read())).convert("RGB")
        # Add image preprocessing (resize, normalize, etc.)
        img_tensor = preprocess_image(img)  # Implement this function
    else:
        img_tensor = None

    # Process text
    input_ids, attention_mask = tokenize_text(text)  # Implement tokenizer

    # Run model
    with torch.no_grad():
        outputs = model(img_tensor, input_ids, attention_mask, task=task)
    
    # Format results based on task
    result = format_results(outputs, task)  # Implement based on task type
    return jsonify(result)