requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
timm>=0.9.0
wandb>=0.15.0
tqdm>=4.65.0
ml-collections==0.1.1  # Pinned to a version compatible with Python 3.9
Pillow>=9.5.0

# Agent dependencies
ollama>=0.2.0
duckduckgo-search>=5.0.0